melee: added a post-parry delay, fixed enemy position prediction save/load bug
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@3464 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ DarkMod/MeleeWeapon.cpp
@@ -1000,7 +1000,7 @@ void CMeleeWeapon::AttachedToActor(idActor *actor)
 		// overwrite AI's initial attack prediction time if we have this spawnarg
 		const idKeyValue *kv = spawnArgs.FindKey("predicted_attack_time");
 		if( kv )
-			actor->m_MeleePredictedAttTime = 0.001f * (float) atoi(kv->GetValue().c_str());
+			actor->m_MeleePredictedAttTime = 0.001f * atof(kv->GetValue().c_str());
 	}
 
 }
